Changes made:
- Added tasks
- More robust logging and error handling introduced across entire backend platform
- Removed log_http_requests, using built in Flask request decorators
- Added status page
- Fixed blank glucose log submission bug

+++ backend/routes/auth_routes.py
@@ -9,8 +9,8 @@ import time
 import hashlib
 from sqlalchemy.exc import SQLAlchemyError
 from utils.db_module import db
-from utils.logger import route_logger
-from utils.utils import handle_sqlalchemy_errors, handle_request_errors, log_http_requests
+from utils.logger import route_logger, error_logger
+from utils.utils import handle_sqlalchemy_errors, handle_request_errors
 from models.user_management_models import Account, User, RefreshToken
 
 # Register auth_routes as a blueprint for importing into app.py + set up CORS
@@ -31,59 +31,67 @@ CORS(auth_routes, supports_credentials=True)
 """
 @auth_routes.route('/login', methods=['POST'])
 @handle_sqlalchemy_errors
-@log_http_requests
 def login():
-    expected_fields = ['email', 'password']
-    data, error = handle_request_errors(request, expected_fields)
-    if error:
-        return error, 400
+    request_id = str(uuid.uuid4())
+    try:
+        expected_fields = ['email', 'password']
+        data, error = handle_request_errors(request, expected_fields)
+        if error:
+            route_logger.info(f"Request {request_id}: Missing fields in request data: {data}")
+            return jsonify(message="Missing required fields"), 400
 
-    email = data['email'].strip()
-    hashed_password = hashlib.sha3_256(data['password'].encode()).digest() 
-
-    account = Account.query.filter_by(email=email).first()
-    
-    if not account:
-        time.sleep(0.1)
-        route_logger.info(f"Attempt by email: '{email}' to log in, no matching account.")
-        return jsonify(message = 'Could not log in with provided credentials. Please try again.'), 401
-    
-    if hashed_password == account.password_hash:
-        account.last_login_date = datetime.now().strftime("%Y:%m:%d %H:%M:%S")
-        access_token_payload = {
-            'user_id': account.id,
-            'username': account.username,
-            'is_admin': account.is_admin
-        }
-        refresh_token_payload = {
-            'user_id': account.id
-        }
-
-        access_token = create_access_token(identity=access_token_payload)
-        refresh_token = create_refresh_token(identity=refresh_token_payload)
+        email = data['email'].strip()
+        hashed_password = hashlib.sha3_256(data['password'].encode()).digest()
         
-        expiration_time = datetime.now() + timedelta(days=7)
-        session_id = str(uuid.uuid4())
-        new_refresh_token = RefreshToken(user_id=account.id, refresh_token=refresh_token, expiration_time=expiration_time, session_id=session_id)
-        db.session.add(new_refresh_token)
-        db.session.commit()
+        route_logger.info(f"Request {request_id}: Attempting login for email: {email}")
 
-        response_data = {
-            'access_token': access_token,
-            'refresh_token': refresh_token,
-            'session_id': session_id
-        }
+        account = Account.query.filter_by(email=email).first()
+
+        if not account:
+            time.sleep(0.1)
+            route_logger.info(f"Request {request_id}: No matching account for email: '{email}'")
+            return jsonify(message='Could not log in with provided credentials. Please try again.'), 401
+
+        if hashed_password == account.password_hash:
+            account.last_login_date = datetime.now().strftime("%Y:%m:%d %H:%M:%S")
+            access_token_payload = {
+                'user_id': account.id,
+                'username': account.username,
+                'is_admin': account.is_admin
+            }
+            refresh_token_payload = {
+                'user_id': account.id
+            }
+
+            access_token = create_access_token(identity=access_token_payload)
+            refresh_token = create_refresh_token(identity=refresh_token_payload)
+            
+            expiration_time = datetime.now() + timedelta(days=7)
+            session_id = str(uuid.uuid4())
+            new_refresh_token = RefreshToken(user_id=account.id, refresh_token=refresh_token, expiration_time=expiration_time, session_id=session_id)
+            db.session.add(new_refresh_token)
+            db.session.commit()
+
+            response_data = {
+                'access_token': access_token,
+                'refresh_token': refresh_token,
+                'session_id': session_id
+            }
+            
+            user_type = "admin" if account.is_admin == 1 else "user"
+            route_logger.info(f"Request {request_id}: {user_type} '{account.username}' logged in successfully. Session ID: {session_id}")
+            return jsonify(response_data), 200
         
-        userType = "admin" if account.is_admin == 1 else "user"
-        route_logger.info(f"{userType} '{account.username}' logged in successfully.")
-        return jsonify(response_data), 200
-    
-    else:
-        account.failed_logins += 1
-        db.session.commit()
-        route_logger.info(f"{account.username}' failed to log in successfully: bad credentials")
-        time.sleep(0.1)
-        return jsonify(message = 'Could not log in with provided credentials. Please try again.'), 401
+        else:
+            account.failed_logins += 1
+            db.session.commit()
+            route_logger.info(f"Request {request_id}: '{account.username}' failed to log in: bad credentials")
+            time.sleep(0.1)
+            return jsonify(message='Could not log in with provided credentials. Please try again.'), 401
+    except Exception as e:
+        error_message = f"Request {request_id}: Error during login process: {str(e)}"
+        error_logger.error(error_message)
+        return jsonify(message="Internal Server Error"), 500
 
 """
     /auth/logout API endpoint:
@@ -98,21 +106,35 @@ def login():
 @auth_routes.route('/logout', methods=['POST'])
 @jwt_required()
 @handle_sqlalchemy_errors
-@log_http_requests
 def logout():
-    access_token = get_jwt_identity()
-    user_id = access_token['user_id']
-    data = request.get_json()
-    session_id = data['session_id']
-    refresh_token = RefreshToken.query.filter_by(user_id=user_id, session_id=session_id).first()
+    request_id = str(uuid.uuid4())
+    try:
+        access_token = get_jwt_identity()
+        user_id = access_token['user_id']
+        route_logger.info(f"Request {request_id}: Received logout request for user ID: {user_id}")
+        
+        data = request.get_json()
+        if not data or 'session_id' not in data:
+            route_logger.info(f"Request {request_id}: Missing session_id in request data")
+            return jsonify(message="Missing required fields"), 400
 
+        session_id = data['session_id']
+        route_logger.info(f"Request {request_id}: Processing logout for session ID: {session_id}")
+        
+        refresh_token = RefreshToken.query.filter_by(user_id=user_id, session_id=session_id).first()
 
-    if refresh_token:
-        db.session.delete(refresh_token)
-        db.session.commit()
-        return jsonify(message = "Logout successful"), 200
-    else:
-        return jsonify(message="Invalid authentication"), 401
+        if refresh_token:
+            db.session.delete(refresh_token)
+            db.session.commit()
+            route_logger.info(f"Request {request_id}: Logout successful for user ID: {user_id}, session ID: {session_id}")
+            return jsonify(message="Logout successful"), 200
+        else:
+            route_logger.info(f"Request {request_id}: Invalid session ID: {session_id} for user ID: {user_id}")
+            return jsonify(message="Invalid authentication"), 401
+    except Exception as e:
+        error_message = f"Request {request_id}: Error during logout process: {str(e)}"
+        error_logger.error(error_message)
+        return jsonify(message="Internal Server Error"), 500
 
 """
     /auth/refresh API endpoint:
@@ -123,31 +145,50 @@ def logout():
 """
 @auth_routes.route('/refresh', methods=['POST'])
 @handle_sqlalchemy_errors
-@log_http_requests
 def refresh():
-    refresh_token = request.cookies.get('refresh_token')
-    session_id = request.cookies.get('session_id')
+    request_id = str(uuid.uuid4())
+    try:
+        data = request.get_json()
+        refresh_token = data['refresh_token']
+        session_id = data['session_id']
 
-    user_id = decode_token(refresh_token)
+        if not refresh_token or not session_id:
+            route_logger.info(f"Request {request_id}: Missing refresh_token or session_id in cookies")
+            return jsonify(message="Missing required cookies"), 401
 
-    stored_token = RefreshToken.query.filter_by(user_id=user_id, session_id=session_id, refresh_token=refresh_token).first()
-    if not stored_token:
-        return jsonify(message="Invalid session credentials."), 401
-    
-    refresh_token_expiration = datetime.strptime(stored_token.expiration_time, '%Y-%m-%d %H:%M:%S.%f')
-    current_time = datetime.now() 
-    if current_time > refresh_token_expiration:
-        return jsonify(message="Expired refresh privilege credentials. Log in again."), 401
-    
-    account = Account.query.get(user_id)
-    access_token_payload = {
-        'user_id': account.id,
-        'username': account.username,
-        'is_admin': account.is_admin
-    }
-    
-    new_access_token = create_access_token(identity=access_token_payload)
-    return jsonify(access_token=new_access_token), 200
+        user_id = decode_token(refresh_token).get('identity')
+        if not user_id:
+            route_logger.info(f"Request {request_id}: Invalid refresh token")
+            return jsonify(message="Invalid refresh token"), 401
+
+        route_logger.info(f"Request {request_id}: Refresh request for user ID: {user_id} with session ID: {session_id}")
+
+        stored_token = RefreshToken.query.filter_by(user_id=user_id, session_id=session_id, refresh_token=refresh_token).first()
+        if not stored_token:
+            route_logger.info(f"Request {request_id}: Invalid session credentials for user ID: {user_id}")
+            return jsonify(message="Invalid session credentials."), 401
+
+        refresh_token_expiration = datetime.strptime(stored_token.expiration_time, '%Y-%m-%d %H:%M:%S.%f')
+        current_time = datetime.now()
+        if current_time > refresh_token_expiration:
+            route_logger.info(f"Request {request_id}: Expired refresh token for user ID: {user_id}")
+            return jsonify(message="Expired refresh privilege credentials. Log in again."), 401
+
+        account = Account.query.get(user_id)
+        access_token_payload = {
+            'user_id': account.id,
+            'username': account.username,
+            'is_admin': account.is_admin
+        }
+
+        new_access_token = create_access_token(identity=access_token_payload)
+        route_logger.info(f"Request {request_id}: Successfully generated new access token for user ID: {user_id}")
+
+        return jsonify(access_token=new_access_token), 200
+    except Exception as e:
+        error_message = f"Request {request_id}: Error during token refresh process: {str(e)}"
+        error_logger.error(error_message)
+        return jsonify(message="Internal Server Error"), 500
 
 """
     /auth/forgot-password API endpoint:
@@ -157,21 +198,29 @@ def refresh():
         - Return 400 on badly formatted email messages
 """
 @auth_routes.route('/forgot-password', methods=['POST'])    
-@log_http_requests
 def forgot_password():
-    data = request.get_json()
-    email = data.get('email', '').strip()
-    
-    if not re.match(r"[^@]+@[^@]+\.[^@]+", email):
-        return jsonify({'message': 'Please enter a valid email.',}), 400
-    
-    #TODO: need method for sending email and recovery options
+    request_id = str(uuid.uuid4())
+    try:
+        data = request.get_json()
+        email = data.get('email', '').strip()
+        
+        if not re.match(r"[^@]+@[^@]+\.[^@]+", email):
+            route_logger.info(f"Request {request_id}: Invalid email format provided: '{email}'")
+            return jsonify({'message': 'Please enter a valid email.'}), 400
+        
+        # TODO: Add method for sending email and recovery options
+        route_logger.info(f"Request {request_id}: Password recovery email requested for: '{email}'")
 
-    response_data = {
-        'message': 'Thank you! Please check your email for recovery options.',
-    }
-    
-    return jsonify(response_data), 200
+        response_data = {
+            'message': 'Thank you! Please check your email for recovery options.',
+        }
+        
+        return jsonify(response_data), 200
+
+    except Exception as e:
+        error_message = f"Request {request_id}: Unexpected error during password recovery: {str(e)}"
+        error_logger.error(error_message)
+        return jsonify({'message': 'Server error occurred.'}), 500
 
 """
     /auth/register API endpoint:
@@ -185,64 +234,48 @@ def forgot_password():
             - 500: Database issues
 """
 @auth_routes.route('/register', methods=['POST'])
-@log_http_requests
+@handle_sqlalchemy_errors
 def register():
-    data = request.get_json()
-
-    required_fields = ['first_name', 'last_name', 'email', 'username', 'password']
-    if not all(field in data for field in required_fields):
-        response_data = {
-            'message': 'Missing fields! Please provide all required fields.',
-        }
-        return jsonify(response_data), 400
-    
-    first_name = data.get('first_name').strip()
-    last_name = data.get('last_name').strip()
-    email = data.get('email').strip()
-    username = data.get('username').strip()
-    password = data.get('password', '').strip()
-    hashed_password = hashlib.sha3_256(password.encode()).digest()
-
-    if not re.match(r"[^@]+@[^@]+\.[^@]+", email):
-        response_data = {
-            'message': 'Incorrect email format. Please provide a proper email.',
-        }
-        return jsonify(response_data), 400
-    
-    if Account.query.filter_by(email=data['email']).first() or Account.query.filter_by(username=data['username']).first():
-        response_data = {
-            'message': 'A user already exists with the provided email or username.',
-        }
-        return jsonify(response_data), 401
-
+    request_id = str(uuid.uuid4())
     try:
-        new_account = Account(
-            username=username,
-            email=email,
-            password_hash=hashed_password
-        )
+        data = request.get_json()
+
+        required_fields = ['first_name', 'last_name', 'email', 'username', 'password']
+        if not all(field in data for field in required_fields):
+            route_logger.info(f"Request {request_id}: Missing fields in registration data")
+            return jsonify(message='Missing fields! Please provide all required fields.'), 400
+
+        first_name = data.get('first_name', '').strip()
+        last_name = data.get('last_name', '').strip()
+        email = data.get('email', '').strip()
+        username = data.get('username', '').strip()
+        password = data.get('password', '').strip()
+        hashed_password = hashlib.sha3_256(password.encode()).digest()
+
+        if not re.match(r"[^@]+@[^@]+\.[^@]+", email):
+            route_logger.info(f"Request {request_id}: Incorrect email format: {email}")
+            return jsonify(message='Incorrect email format. Please provide a proper email.'), 400
+
+        existing_email = Account.query.filter_by(email=email).first()
+        existing_username = Account.query.filter_by(username=username).first()
+
+        if existing_email or existing_username:
+            route_logger.info(f"Request {request_id}: Email '{email}' or username '{username}' already exists.")
+            return jsonify(message='A user already exists with the provided email or username.'), 401
+
+        new_account = Account(username=username, email=email, password_hash=hashed_password)
         db.session.add(new_account)
         db.session.commit()
 
-        # Update user table with backreference to the account table, new account now present
         user_id = new_account.id
-        new_user = User(
-            id=user_id,
-            first_name=first_name,
-            last_name=last_name,
-        )
+        new_user = User(id=user_id, first_name=first_name, last_name=last_name)
         db.session.add(new_user)
-        db.session.commit() 
-        
-        response_data = {
-            'message': 'Registration successful! Check your email for verification.',
-        }
-        route_logger.info(f"User '{new_account.username}' with id {new_account.id} registered and was added to database.")
-        return jsonify(response_data), 200
+        db.session.commit()
 
-    except SQLAlchemyError:
-        response_data = {
-            'message': 'Server error occurred.',
-        }
-        db.session.rollback()
-        return jsonify(response_data), 500
+        route_logger.info(f"Request {request_id}: User '{new_account.username}' with ID {new_account.id} registered successfully.")
+        return jsonify(message='Registration successful! Check your email for verification.'), 200
+
+    except Exception as e:
+        error_message = f"Request {request_id}: Unexpected error during registration: {str(e)}"
+        error_logger.error(error_message)
+        return jsonify(message='Server error occurred.'), 500
